Default image functionality
- default image for a newRecipe in the recipe schema-  image defaults to url in recipe schema:
In the recipe route for a Post request, checks to see if the image has a length, (ie. isn't an empty string-if you start typing and then get rid of text, the input field will be an empty string), if it is then it is set to undefined, so the default image gets picked up.
- Functionality for when searchValue is equal to "", then display the full recipeList data, so p tag of 'how about adding some recipes' doesn't show